# Conflicts:
#	.gradle/3.3/taskArtifacts/fileHashes.bin
#	.gradle/3.3/taskArtifacts/fileSnapshots.bin
#	.gradle/3.3/taskArtifacts/taskArtifacts.bin
#	.gradle/3.3/taskArtifacts/taskArtifacts.lock
#
.gradle/3.3/tasks/_app_compileDebugJavaWithJavac/localClassSetAnalysis/l
ocalClassSetAnalysis.bin
#
.gradle/3.3/tasks/_app_compileDebugJavaWithJavac/localClassSetAnalysis/l
ocalClassSetAnalysis.lock
#
.gradle/3.3/tasks/_app_compileDebugJavaWithJavac/localJarClasspathSnapsh
ot/localJarClasspathSnapshot.lock
#	.idea/libraries/SMSSDK_2_1_4_.xml
#	.idea/libraries/appcompat_v7_23_1_1.xml
#	.idea/libraries/cardview_v7_23_1_0.xml
#	.idea/libraries/cardview_v7_23_1_1.xml
#	.idea/libraries/constraint_layout_1_0_2.xml
#	.idea/libraries/support_v4_23_1_1.xml
#	Android_RunForYou.iml
#	build/intermediates/dex-cache/cache.xml
#	local.properties

+++ app/src/main/java/xzh/com/materialdesign/ui/DetailsActivity.java
@@ -87,14 +87,14 @@ public class DetailsActivity extends AppCompatActivity {
         if(ordersInfo != null){
             title.setText(ordersInfo.getOrderItem());
             name.setText(ordersInfo.getContactName());
-            reward.setText(String.valueOf(ordersInfo.getOrderReward()));
+            reward.setText(String.valueOf((int)ordersInfo.getOrderReward()));
 
             if(ordersInfo.getOrderType() == 1){
                 method.setText("分");
             }else
                 method.setText("元");
 
-            money.setText(String.valueOf(ordersInfo.getOrderPredict())+"元");
+            money.setText(String.valueOf((int)ordersInfo.getOrderPredict())+"元");
             info.setText(ordersInfo.getOrderDescribe());
             shop.setText(ordersInfo.getOrderDestination());
             des.setText(ordersInfo.getOrderAddress());
